[#1] Removing Ribbon namespace.
