+++ lib/plugins/version.rb
@@ -1,0 +1,3 @@
+class Plugins
+  VERSION = "0.3.0"
+end
