--- lib/ribbon/plugins/version.rb
@@ -1,5 +1,0 @@
-module Ribbon
-  class Plugins
-    VERSION = "0.2.4"
-  end
-end
