+++ ribbon-plugins.gemspec
@@ -1,16 +1,16 @@
 $:.push File.expand_path("../lib", __FILE__)
 
 # Maintain your gem's version:
-require "ribbon/plugins/version"
+require "plugins/version"
 
 Gem::Specification.new do |s|
-  s.name        = 'ribbon-plugins'
-  s.version     = Ribbon::Plugins::VERSION
-  s.homepage    = "http://github.com/ribbon/plugins"
+  s.name        = 'plugins'
+  s.version     = Plugins::VERSION
+  s.homepage    = "http://github.com/payout/plugins"
   s.summary     = "A flexible plugins framework."
   s.description = s.summary
   s.authors     = ["Robert Honer"]
-  s.email       = ['robert@ribbonpayments.com']
+  s.email       = ['robert@payout.com']
   s.files       = Dir['lib/**/*.rb']
   s.license     = 'BSD'
