+++ spec/around_stack_spec.rb
@@ -1,4 +1,4 @@
-class Ribbon::Plugins
+class Plugins
   RSpec.describe AroundStack do
     let(:stack) { AroundStack.new(:test) }
     subject { stack }
