+++ spec/spec_helper.rb
@@ -1,6 +1,6 @@
 require "codeclimate-test-reporter"
 CodeClimate::TestReporter.start
-require 'ribbon/plugins'
+require 'plugins'
 
 # This file was generated by the `rspec --init` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
